refactor: 관심 모임 나가기 렌더링 컴포넌트 변경
